fix and clarify broken link
The link referenced was renamed to "how to work a challenge"; Several Otters reported that the link was broken last weekend. They also didn't understand that the resource they were trying to find was already in their phase guide.

Link is now synced back up, and I edited the sentence around it to clarify that the resource is one they've already seen.